renamed schema directory
