+++ tests/schema.test.ts
@@ -11,7 +11,7 @@ import request from 'supertest'
 import { setupServer } from '../src/server'
 
 import { getTestAgent, getTestSchema } from './utils/helpers'
-import _schema from '../schemainformation/schemaAttributes.json'
+import _schema from '../schema/schemaAttributes.json'
 const schema = _schema as AnonCredsSchema
 
 describe('SchemaController', () => {
